SNAP-66: Staging people-summary
